Cleanup API endpoints
