+++ backend/src/db-utils.js
@@ -581,23 +581,6 @@ const calculateFeesByDay = async (days = 1) => {
   );
 };
 
-const calculateFeePerDay = async (days = 1) => {
-  return querySingleRow(
-    [
-      `SELECT
-          SUM(transactions.receipt_conversion_tokens_burnt + eo.tokens_burnt) AS fee
-      FROM transactions
-      LEFT OUTER JOIN receipts on receipts.originated_from_transaction_hash = transactions.transaction_hash
-      JOIN execution_outcomes eo on eo.receipt_id = receipts.receipt_id
-      WHERE
-        block_timestamp > (CAST(EXTRACT(EPOCH FROM DATE_TRUNC('day', NOW() - INTERVAL '${days} day')) AS bigint) * 1000 * 1000 * 1000)
-      AND
-      block_timestamp < (CAST(EXTRACT(EPOCH FROM DATE_TRUNC('day', NOW())) AS bigint) * 1000 * 1000 * 1000)`,
-    ],
-    { dataSource: DS_INDEXER_BACKEND }
-  );
-};
-
 // node part
 exports.queryOnlineNodes = queryOnlineNodes;
 exports.extendWithTelemetryInfo = extendWithTelemetryInfo;
@@ -641,4 +624,3 @@ exports.getLastYesterdayBlock = getLastYesterdayBlock;
 
 // calculate fee
 exports.calculateFeesByDay = calculateFeesByDay;
-exports.calculateFeePerDay = calculateFeePerDay;
