+++ backend/src/wamp.js
@@ -322,10 +322,6 @@ wampHandlers["nearcore-total-fee-count"] = async ([daysCount]) => {
   return await stats.getTotalFee(daysCount);
 };
 
-wampHandlers["nearcore-total-fee-count-per-days"] = async ([daysCount]) => {
-  return await stats.getTotalFeePerDays(daysCount);
-};
-
 // set up wamp
 function setupWamp() {
   const wamp = new autobahn.Connection({
