--- frontend/src/pages/api/fee-per-day.js
@@ -1,11 +1,0 @@
-import StatsApi from "../../libraries/explorer-wamp/stats";
-
-export default async function (req, res) {
-  try {
-    const { fee } = await new StatsApi(req).getTotalFeePerDays(1);
-    res.send({ totalFeeInYoctoNear: fee });
-  } catch (error) {
-    console.error(error);
-    res.status(502).send(error);
-  }
-}
